requirements.txt
Django==1.9.9
PyJWT==1.4.2
Pygments==2.1.3
apiclient==1.0.3
argparse==1.2.1
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
djangorestframework==3.4.6
enum34==1.1.6
google-api-python-client==1.5.3
httplib2==0.9.2
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
oauth2client==3.0.0
oauthlib==2.0.0
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.7
psycopg2==2.6.2
ptyprocess==0.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-openid==2.2.5
python-social-auth==0.2.21
requests==2.11.1
requests-oauthlib==0.6.2
rest-social-auth==0.5.0
rsa==3.4.2
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
traitlets==4.3.0
uritemplate==0.6
urllib3==1.17
wcwidth==0.1.7
wsgiref==0.1.2
